requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.3
mysql-connector-python==8.0.33
python-dotenv==1.0.0
Flask-Cors==4.0.1
Flask-Login==0.6.3
Werkzeug==2.3.6
email-validator==2.0.0
